ci: add CI to lint the codebase. (#78)
I've found some inconsistencies in the code. So, perhaps this PR will
solve the issue.

This PR adds:
- [typos by crate-ci](https://github.com/crate-ci/typos), to fix and
check for typos.
- [dprint by dprint](https://github.com/dprint/dprint), to properly
format markdown, YAML, and Dockerfile files.

The mentioned tools are also included inside the `justfile`. So, one is
able to check locally just using `just fmt` or `just fmt-check`.