feat: add admin data explorer. (#208)
Adds a primitive admin explorer to the admin tools:

`http://localhost:9523/__internal__/admin/explorer`


![image](https://github.com/user-attachments/assets/99d650cb-2d33-4d9e-8088-fb80f98c571e)

- **why we needed it:** There was a bug that allowed you to put slashes
in the page slugs, causing a situation where it was impossible to
navigate to the page and delete it. This meant we had to have a way, as
an admin, to go straight to the data and delete the offending page, so
that the user wasn't stuck with the broken page eternally in their
listing. We may have to do other, similar kinds of edits over time, so
this will enable us to take quick action to fix problems instead of
having to produce a migration in code and re-deploy before being able to
fix the issue.
- **what it does:** Adds a new page to the admin tools that allows you
to see all of the entities in the data and read / modify most of the
components that are used by Weird.
- **how it was made:** Added a new RPC endpoint that allowed listing
namespaces and subspaces, and otherwise works similar to the recent
database dump feature to let you browse all of the entities, and then
edit them.