chore: introduce a `.editorconfig` for TypeScript/Svelte (#257)
editorconfig is widely supported in most popular editors today, it helps
writing code in different file extensions without having to update
editor configuration by hand.

This PR provides an editorconfig for TypeScript and Svelte to use tabs
for indentation of size 2 and in general use EOL line feed with newline.